ToSchema for Rational, tests to pin it in place
