+++ plutus-numeric/golden/data/PlutusTx-Rational-Internal-Rational.json
@@ -1,0 +1,285 @@
+{
+    "seed": 123456,
+    "data": [
+        {
+            "data": [
+                {
+                    "data": 29,
+                    "tag": "I"
+                },
+                {
+                    "data": 47,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": -17,
+                    "tag": "I"
+                },
+                {
+                    "data": 5,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": -26,
+                    "tag": "I"
+                },
+                {
+                    "data": 75,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": -8,
+                    "tag": "I"
+                },
+                {
+                    "data": 25,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": -34,
+                    "tag": "I"
+                },
+                {
+                    "data": 25,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": -84,
+                    "tag": "I"
+                },
+                {
+                    "data": 55,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": 11,
+                    "tag": "I"
+                },
+                {
+                    "data": 31,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": 18,
+                    "tag": "I"
+                },
+                {
+                    "data": 47,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": 91,
+                    "tag": "I"
+                },
+                {
+                    "data": 87,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": -89,
+                    "tag": "I"
+                },
+                {
+                    "data": 17,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": -67,
+                    "tag": "I"
+                },
+                {
+                    "data": 3,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": 1,
+                    "tag": "I"
+                },
+                {
+                    "data": 1,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": 81,
+                    "tag": "I"
+                },
+                {
+                    "data": 37,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": -96,
+                    "tag": "I"
+                },
+                {
+                    "data": 67,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": -27,
+                    "tag": "I"
+                },
+                {
+                    "data": 7,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": -19,
+                    "tag": "I"
+                },
+                {
+                    "data": 37,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": -57,
+                    "tag": "I"
+                },
+                {
+                    "data": 100,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": 7,
+                    "tag": "I"
+                },
+                {
+                    "data": 26,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": 55,
+                    "tag": "I"
+                },
+                {
+                    "data": 46,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        },
+        {
+            "data": [
+                {
+                    "data": -7,
+                    "tag": "I"
+                },
+                {
+                    "data": 11,
+                    "tag": "I"
+                }
+            ],
+            "index": 0,
+            "tag": "Constr"
+        }
+    ]
+}
